+++ src/main/java/com/sojoo/StoreSpotter/controller/user/SingInController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
-public class SingIn {
+public class SingInController {
 
     // 로그인
     @GetMapping("/signin")
